initial sketch of a way to do lists
But I'm starting to realize that these things are rarely readable in
isolation. We also need the ability to massage the prose, excerpt it.
All while retaining a permalink for provenance.

Or perhaps lists are a stupid idea. My initial goal here was a place to
share individual links. An assorted list of favorites is not very
useful out of context.